+++ spec/www/spec/regexp-test.js
@@ -1,0 +1,76 @@
+/* 'use strict'; */
+
+var MYTIMEOUT = 12000;
+
+var DEFAULT_SIZE = 5000000; // max to avoid popup in safari/ios
+
+var isAndroid = /Android/.test(navigator.userAgent);
+var isWindows = /Windows /.test(navigator.userAgent); // Windows (8.1)
+
+var scenarioList = [ isAndroid ? 'Plugin-sqlite-connector' : 'Plugin', 'HTML5', 'Plugin-android.database' ];
+
+var scenarioCount = isAndroid ? 3 : (isIE ? 1 : 2);
+
+// simple tests:
+var mytests = function() {
+
+  for (var i=0; i<scenarioCount; ++i) {
+
+    describe(scenarioList[i] + ': REGEX test(s)', function() {
+      var scenarioName = scenarioList[i];
+      var suiteName = scenarioName + ': ';
+      var isWebSql = (i === 1);
+      var isOldImpl = (i === 2);
+
+      // NOTE: MUST be defined in function scope, NOT outer scope:
+      var openDatabase = function(name, ignored1, ignored2, ignored3) {
+        if (isOldImpl) {
+          return window.sqlitePlugin.openDatabase({name: name, androidDatabaseImplementation: 2});
+        }
+        if (isWebSql) {
+          return window.openDatabase(name, "1.0", "Demo", DEFAULT_SIZE);
+        } else {
+          return window.sqlitePlugin.openDatabase(name, "1.0", "Demo", DEFAULT_SIZE);
+        }
+      }
+
+      it(suiteName + 'Simple REGEXP test',
+        function(done) {
+          if (isWebSql && !isAndroid) pending('BROKEN for iOS Web SQL');
+          if (!isWebSql && isAndroid && !isOldImpl) pending('BROKEN for default Android implementation');
+          if (isWindows) pending('BROKEN for Windows ("Universal")');
+
+          var db = openDatabase('simple-regexp-test.db', '1.0', 'test', DEFAULT_SIZE);
+
+          expect(db).toBeDefined();
+
+          db.transaction(function(tx) {
+
+            expect(tx).toBeDefined();
+            tx.executeSql('DROP TABLE IF EXISTS tt');
+            tx.executeSql('CREATE TABLE tt (tv TEXT)');
+
+            tx.executeSql('INSERT INTO tt VALUES (?)', ['test']);
+            tx.executeSql('INSERT INTO tt VALUES (?)', ['tst2']);
+
+            tx.executeSql("SELECT * from tt WHERE tv REGEXP('te?st2+')", [], function(tx, res) {
+              expect(res.rows.length).toBe(1);
+              expect(res.rows.item(0).tv).toBe('tst2');
+
+              done();
+            }, function(e) {
+              // went wrong:
+              expect(false).toBe(true);
+              done();
+            });
+          });
+        }, MYTIMEOUT);
+
+    });
+  };
+}
+
+if (window.hasBrowser) mytests();
+else exports.defineAutoTests = mytests;
+
+/* vim: set expandtab : */