+++ src/ios/SQLitePlugin.m
@@ -141,11 +141,11 @@ static void sqlite_regexp(sqlite3_context* context, int argc, sqlite3_value** va
 
             NSLog(@"open full db path: %@", dbname);
 
-            /* Option to create from resource (pre-populated) if db does not exist: */
-            if (![[NSFileManager defaultManager] fileExistsAtPath:dbname]) {
-                NSString *createFromResource = [options objectForKey:@"createFromResource"];
+            /* Option to create database from resource (pre-populated) if it does not exist: */
+            if (![[NSFileManager defaultManager] fileExistsAtPath: dbname]) {
+                NSString * createFromResource = [options objectForKey:@"createFromResource"];
                 if (createFromResource != NULL)
-                    [self createFromResource:dbfilename withDbname:dbname];
+                    [self createFromResource: dbfilename withDbname: dbname];
             }
 
             if (sqlite3_open(name, &db) != SQLITE_OK) {
@@ -186,21 +186,21 @@ static void sqlite_regexp(sqlite3_context* context, int argc, sqlite3_value** va
 }
 
 
--(void)createFromResource:(NSString *)dbfile withDbname:(NSString *)dbname {
-    NSString *bundleRoot = [[NSBundle mainBundle] resourcePath];
-    NSString *www = [bundleRoot stringByAppendingPathComponent:@"www"];
-    NSString *prepopulatedDb = [www stringByAppendingPathComponent: dbfile];
-    // NSLog(@"Look for prepopulated DB at: %@", prepopulatedDb);
+-(void) createFromResource: (NSString *)dbfile withDbname:(NSString *)dbname {
+    NSString * bundleRoot = [[NSBundle mainBundle] resourcePath];
+    NSString * www = [bundleRoot stringByAppendingPathComponent:@"www"];
+    NSString * prepopulatedDb = [www stringByAppendingPathComponent: dbfile];
+    // NSLog(@"Look for pre-populated DB at: %@", prepopulatedDb);
 
     if ([[NSFileManager defaultManager] fileExistsAtPath:prepopulatedDb]) {
         NSLog(@"Found prepopulated DB: %@", prepopulatedDb);
-        NSError *error;
+        NSError * error;
         BOOL success = [[NSFileManager defaultManager] copyItemAtPath:prepopulatedDb toPath:dbname error:&error];
 
         if(success)
-            NSLog(@"Copied prepopulated DB content to: %@", dbname);
+            NSLog(@"Copied pre-populated DB content to: %@", dbname);
         else
-            NSLog(@"Unable to copy DB file: %@", [error localizedDescription]);
+            NSLog(@"Unable to copy pre-populated DB file: %@", [error localizedDescription]);
     }
 }