+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 0.8.0-ext-dev
 
+- Support pre-populated database for Windows "Universal"
 - Fix conversion warnings in iOS version
 - Fix to Windows "Universal" version to support big integers
 - Implement database close and delete operations for Windows "Universal"